+++ src/main/java/com/wanted/socialMediaIntegratedFeed/domain/post/PostRepositoryCustom.java
@@ -1,10 +1,8 @@
 package com.wanted.socialMediaIntegratedFeed.domain.post;
 
-import com.wanted.socialMediaIntegratedFeed.api.content.dto.response.PostPaginationResponse;
+import com.wanted.socialMediaIntegratedFeed.web.content.dto.response.PostPaginationResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-
-import java.util.List;
 
 public interface PostRepositoryCustom {
     Page<PostPaginationResponse> findAllByHashtag(String hashtag,String type,String searchBy,String search,Pageable pageable);
