+++ src/main/java/com/wanted/socialMediaIntegratedFeed/web/content/service/PostServiceImpl.java
@@ -1,6 +1,6 @@
-package com.wanted.socialMediaIntegratedFeed.api.content.service;
+package com.wanted.socialMediaIntegratedFeed.web.content.service;
 
-import com.wanted.socialMediaIntegratedFeed.api.content.dto.response.PostPaginationResponse;
+import com.wanted.socialMediaIntegratedFeed.web.content.dto.response.PostPaginationResponse;
 
 import com.wanted.socialMediaIntegratedFeed.domain.post.PostRepository;
